Cap NUMERIC DIGITS, and stop two casts wrapping above it
`Settings` accepted any `u32` as `NUMERIC DIGITS`. The interpreter caps it at
999999999 and raises error 26 above that: 999999999 is accepted and computes
`1/3` to 1000000001 digits, while 1000000000, 2147483647 and 4294967296 are
all rejected. `MAX_EXPONENT` already held that value and `NotWholeNumber`
already carried code 26, so neither a new constant nor a new variant was
needed. `FUZZ` is capped transitively by the existing `fuzz < digits` rule.

Two casts were reachable past it and stay fixed on their own account, since
both functions are `pub` and take a bare `u32` that no longer has to come
from `Settings`:

- `Number::format` doubled `digits` after narrowing to `i32`, overflowing
  above 1073741823. In debug that panics; in release it wraps and silently
  picks the wrong display form.
- `pow`'s `as_whole` compared against `digits as i32`, which wraps negative
  above `i32::MAX` and inverts the test rather than panicking -- wrong
  answer, quietly, in both profiles.

Both now compare in `i64` and are correct across the whole `u32` domain.
Neither clamps: clamping would trade a panic for a wrong answer whenever a
caller legitimately passed something larger.

Found by asking whether the cast defect a reviewer found in `format.rs` had
siblings elsewhere. It had two, in tasks already marked complete.

134 tests pass, clippy clean, and seven differential sets stay at 0 --
49,544 cases across format, power, multiply/divide and add/subtract.

+++ rust/crates/rexx-num/src/pow.rs
@@ -38,7 +38,10 @@ impl Number {
         if self_.is_zero() {
             return Some(0);
         }
-        if self_.digits.len() as i32 + self_.exponent > digits as i32 {
+        // Compared in i64: `digits` narrowed to i32 wraps negative above
+        // i32::MAX, which would invert this comparison rather than panic --
+        // a silently wrong answer, in release and debug alike.
+        if self_.digits.len() as i64 + self_.exponent as i64 > digits as i64 {
             return None;
         }
         if self_.exponent < 0 {
